requirements.txt
# Web framework
fastapi>=0.111.0
uvicorn[standard]>=0.30.0
python-multipart>=0.0.9

# Database
sqlalchemy[asyncio]>=2.0.30
aiosqlite>=0.20.0
duckdb>=0.10.3
pandas>=2.2.0
numpy>=1.26.0
faker>=25.0.0

# Auth
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4

# LLM
groq>=0.9.0
openai>=1.35.0

# Cache
cachetools>=5.3.3
# aioredis>=2.0.1  # uncomment for Redis

# Export
reportlab>=4.2.0
python-docx>=1.1.2
python-pptx>=0.6.23

# Config
pydantic-settings>=2.3.0
python-dotenv>=1.0.1
httpx>=0.27.0

# Testing
pytest>=8.2.0
pytest-asyncio>=0.23.7
